backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
python-dotenv
sqlalchemy
requests
PyJWT
bcrypt